% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hep_search.r
\name{hep_search}
\alias{hep_search}
\title{Get INSPIRE HEP publications}
\usage{
hep_search(p = NULL, jrec = 1, batch_size = 10, limit = 100)
}
\arguments{
\item{p}{search pattern (character vector). INSPIRE HEP supports different 
search syntaxes, including the SPIRES-style. See 
\url{https://inspirehep.net/info/hep/search-tips} for a comprehensive 
overview and examples on how to search HEP literature.}

\item{jrec}{Navigate to a record in the search results. By default, retrieval
starts with 1.}

\item{batch_size}{number of records to be returned for each result page. By 
default, groups 10 records for each page. The maximum number is 250.}

\item{limit}{Maximum number of records (default is 100). Since the API is 
still under development, be respectful and avoid expensive queries. INSPIRE
allows bulk downloads via its OAI-PMH interface 
(\url{https://inspirehep.net/oai2d?verb=Identify}) or through a periodically 
updated json dump (\url{https://inspirehep.net/hep_records.json.gz}).}
}
\value{
The result is a data.frame with key metadata. If you wish to retrieve
  more details for each record see.
  
The \code{data.frame} has the following columns.
\tabular{rll}{
 [,1] \tab id               \tab INSPIRE HEP ID \cr
 [,2] \tab title            \tab title of work \cr
 [,3] \tab author           \tab first author \cr
 [,4] \tab affiliation      \tab first author affiliation (collapsed ";") \cr
 [,5] \tab doi              \tab Digital Object Identifier (DOI) \cr
 [,6] \tab report_number    \tab eprint id from the arXiv (collapsed ";") \cr
 [,7] \tab date_reported    \tab date of initial appearance of preprint \cr
 [,8] \tab journal          \tab journal title \cr
 [,9] \tab volume           \tab journal volume \cr
[,10] \tab issue            \tab journal issue \cr
[,11] \tab keywords         \tab controlled keywords (collapsed ";") \cr
[,12] \tab collection       \tab collection information (collapsed ";") \cr
[,13] \tab license_url      \tab re-use terms for full texts (e.g. CC) \cr
}
}
\description{
Search INSPIRE HEP, an open digital library for literature and data in the 
field of high energy physics.
}
\examples{
\dontrun{
# Taken from the INSPIRE search guide 
# (https://inspirehep.net/info/hep/search-tips)
 
# keywords
hep_search('witten black hole', limit = 5)

# work by exact author
hep_search('exactauthor:D.J.Schwarz.1', limit = 5)

#Full text search available for arXiv eprints
hep_search('find ft "faster than light', limit = 5)

# Find caption
hep_search('find caption "Diagram for the fermion flow violating process"')

# Find eprints
hep_search('find eprint arxiv:0711.2908 or arxiv:0705.4298')

# Navigating the search results
# Get record positions 20 - 29
hep_search('witten black hole', jrec = 20, limit = 10)
}

}
\seealso{
\url{https://inspirehep.net/info/hep/search-tips}
}

